+++ examples/class/logging.rb
@@ -21,6 +21,7 @@ module Node
   module JSON
     class Parser
       include CallableTree::Node::Internal
+      prepend CallableTree::Node::Hooks::Matcher
 
       def match?(input, **_options)
         File.extname(input) == '.json'
@@ -40,6 +41,8 @@ module Node
 
     class Scraper
       include CallableTree::Node::External
+      prepend CallableTree::Node::Hooks::Matcher
+      prepend CallableTree::Node::Hooks::Caller
 
       def initialize(type:)
         @type = type
@@ -64,6 +67,7 @@ module Node
   module XML
     class Parser
       include CallableTree::Node::Internal
+      prepend CallableTree::Node::Hooks::Matcher
 
       def match?(input, **_options)
         File.extname(input) == '.xml'
@@ -82,6 +86,8 @@ module Node
 
     class Scraper
       include CallableTree::Node::External
+      prepend CallableTree::Node::Hooks::Matcher
+      prepend CallableTree::Node::Hooks::Caller
 
       def initialize(type:)
         @type = type
@@ -106,18 +112,50 @@ module Node
   end
 end
 
+module Logging
+  INDENT_SIZE = 2
+  BLANK = ' '
+  LIST_STYLE = '*'
+  INPUT_LABEL  = 'Input :'
+  OUTPUT_LABEL = 'Output:'
+
+  def self.loggable(node)
+    node.after_matcher! do |matched, _node_:, **|
+      prefix = LIST_STYLE.rjust((_node_.depth * INDENT_SIZE) - INDENT_SIZE + LIST_STYLE.length, BLANK)
+      puts "#{prefix} #{_node_.identity}: [matched: #{matched}]"
+      matched
+    end
+
+    if node.external?
+      node
+        .before_caller! do |input, *, _node_:, **|
+          input_prefix = INPUT_LABEL.rjust((_node_.depth * INDENT_SIZE) + INPUT_LABEL.length, BLANK)
+          puts "#{input_prefix} #{input}"
+          input
+        end
+        .after_caller! do |output, _node_:, **|
+          output_prefix = OUTPUT_LABEL.rjust((_node_.depth * INDENT_SIZE) + OUTPUT_LABEL.length, BLANK)
+          puts "#{output_prefix} #{output}"
+          output
+        end
+    end
+  end
+end
+
+loggable = Logging.method(:loggable)
+
 tree = CallableTree::Node::Root.new.append(
-  Node::JSON::Parser.new.append(
-    Node::JSON::Scraper.new(type: :animals).verbosify,
-    Node::JSON::Scraper.new(type: :fruits).verbosify
+  Node::JSON::Parser.new.tap(&loggable).append(
+    Node::JSON::Scraper.new(type: :animals).tap(&loggable).verbosify,
+    Node::JSON::Scraper.new(type: :fruits).tap(&loggable).verbosify
   ),
-  Node::XML::Parser.new.append(
-    Node::XML::Scraper.new(type: :animals).verbosify,
-    Node::XML::Scraper.new(type: :fruits).verbosify
+  Node::XML::Parser.new.tap(&loggable).append(
+    Node::XML::Scraper.new(type: :animals).tap(&loggable).verbosify,
+    Node::XML::Scraper.new(type: :fruits).tap(&loggable).verbosify
   )
 )
 
-Dir.glob("#{__dir__}/docs/*") do |file|
+Dir.glob("#{__dir__}/../docs/*") do |file|
   options = { foo: :bar }
   pp tree.call(file, **options)
   puts '---'